同步3.14的tiledmap模块，支持最新的tiledmap编辑器参数。注：不建议使用fastTMX模块。 CCImage加入getColorAt接口。

+++ cocos/platform/CCFileUtils.cpp
@@ -306,7 +306,7 @@ public:
         _state = SAX_NONE;
     }
 
-    void textHandler(void *ctx, const char *ch, int len)
+    void textHandler(void *ctx, const char *ch, size_t len)
     {
         CC_UNUSED_PARAM(ctx);
         if (_state == SAX_NONE)
