+++ cocos/platform/CCSAXParser.h
@@ -23,6 +23,7 @@
 
 #ifndef __CCSAXPARSER_H__
 #define __CCSAXPARSER_H__
+/// @cond DO_NOT_SHOW
 
 #include "platform/CCPlatformConfig.h"
 #include "platform/CCCommon.h"
@@ -56,7 +57,7 @@ public:
      * @js NA
      * @lua NA
      */
-    virtual void textHandler(void *ctx, const char *s, int len) = 0;
+    virtual void textHandler(void *ctx, const char *s, size_t len) = 0;
 };
 
 class CC_DLL SAXParser
@@ -107,7 +108,7 @@ public:
      * @js NA
      * @lua NA
      */
-    static void textHandler(void *ctx, const CC_XML_CHAR *name, int len);
+    static void textHandler(void *ctx, const CC_XML_CHAR *name, size_t len);
 };
 
 // end of platform group
@@ -115,4 +116,5 @@ public:
 
 NS_CC_END
 
+/// @endcond
 #endif //__CCSAXPARSER_H__
